Refactored expression test suites.
